Started working on fixed point numbers.

+++ vtypes.h
@@ -2,6 +2,6 @@
 #define TYPES_H
 
 typedef unsigned char (vbyte);
-typedef unsigned int (vuint);
+typedef unsigned short int (vuint);
 
 #endif
